把出站文件拆成自己的模块，让两半各说自己的话
files.ts 一直担着两件事：字节进来和字节出去。模块注释就是两段拼起来的——后半
段讲的是 ADR 0004 那个"走 Buffer 绕过了 SDK 的路径防护，所以这个顺序是补回来
的"故事，和入站落盘毫无关系。缝就在 OutboundRefusal 开始的地方。

files.ts 留下入站：消毒、落盘、配额、注记；src/outbound-file.ts 拿走校验、读取、
两种拒绝文案、send_file 与 /get。各自的 @module 只讲自己那一件事，ADR 0004 的
来龙去脉归到出站那份。tests/files.spec.ts 同样对半分，导入方一起改。

纯搬移：移动的代码逐字节相同，没有一行逻辑变化。
