+++ tests/plugin.spec.ts
@@ -7,7 +7,7 @@ import Loader from '@deepseek-ai/cordis-plugin-loader'
 import { Context } from '@deepseek-ai/cordis'
 import type { CardActionEvent, NormalizedMessage, SendOptions } from '@larksuite/channel'
 import * as plugin from '../src/index.ts'
-import { GET_COMMAND, SEND_FILE_TOOL } from '../src/files.ts'
+import { GET_COMMAND, SEND_FILE_TOOL } from '../src/outbound-file.ts'
 import { QUESTION_TIMEOUT_MS } from '../src/questions.ts'
 import { parseRoute } from '../src/model.ts'
 import * as invariant from '../src/invariant.ts'
